Reorganizing
- updated test data (use single quote yyyy-mm-dd for dates)
- moved db calls into their own files
- moved app/layout.server.ts into page.server instead, don't need to run
  all that on every load
  - TODO: redirect if the user attempts a non-index url without an id or
    code
- fixed project_name column (was just "name") in project creation query
