+++ src/routes/app/+page.svelte
@@ -2,7 +2,7 @@
 	export let data;
 </script>
 
-<h2 color={data.primaryColor}>Pika!</h2>
+<h2 color={'blue'}>Pika!</h2>
 
 <h3>{data.uuid}</h3>
