+++ src/routes/app/onboard/+page.server.ts
@@ -1,3 +1,4 @@
+import { createProjectByNameAndId } from '$lib/pg/projects.js';
 import { UUID_STRING } from '$lib/stringsAndStuff.js';
 import { redirect } from '@sveltejs/kit';
 
@@ -11,22 +12,13 @@ export const actions = {
 			console.log(
 				`onboard action received request to create project with name ${projectName} for user with id ${uuid} `
 			);
-			let redirectToApp = false;
-			try {
-				const res = await event.locals.pg.query(
-					'insert into projects (user_id, name) values ($1, $2) returning *',
-					[uuid, projectName]
+
+			const res = await createProjectByNameAndId(event.locals.pg, projectName.toString(), uuid);
+
+			if (res?.rows[0].user_id === uuid) {
+				console.log(
+					`successfully added project ${projectName} with id ${res.rows[0].id} for user with id ${res.rows[0].user_id}`
 				);
-				if (res.rows[0].user_id === uuid) {
-					console.log(
-						`successfully added project ${projectName} with id ${res.rows[0].id} for user with id ${res.rows[0].user_id}`
-					);
-					redirectToApp = true;
-				}
-			} catch (err) {
-				console.error(err);
-			}
-			if (redirectToApp) {
 				redirect(303, '/app');
 			}
 		}
